requirements.txt
flask==3.0.2
Flask-Cors==4.0.0
greenlet==3.0.3
groq==0.9.0
numpy==1.24.4
openai==1.14.0
openai-whisper==20231117
PyAudio==0.2.14
pydantic-core==2.16.3
python-decouple==3.8
requests==2.31.0
SpeechRecognition==3.10.1
werkzeug==3.0.1
